Crate chain_listener

Structs

CIDV1

ChainListener

CommitmentActivated

@dev Emitted when a commitment is activated. Commitment can be activated only if delegator deposited collateral. @param peerId Peer id which linked to the commitment @param commitmentId Commitment id which activated @param startEpoch The start epoch of the commitment @param endEpoch The end epoch of the commitment @param unitIds Compute unit ids which linked to the commitment Event with signature CommitmentActivated(bytes32,bytes32,uint256,uint256,bytes32[]) and selector 0x0b0a4688a90d1b24732d05ddf4925af69f02cd7d9a921b1cdcd4a7c2b6d57d68.

ComputeUnitsMatched

Event with signature ComputeUnitsMatched(bytes32,address,bytes32,bytes32[],(bytes4,bytes32)) and selector 0x6e5629a2cfaa82d1ea7ad51936794f666271fbd5068017020eaaafdbb017e615.

UnitActivated

@dev Emitted when a unit activated. Unit is activated when it returned from deal @param commitmentId Commitment id @param unitId Compute unit id which activated Event with signature UnitActivated(bytes32,bytes32,uint256) and selector 0x8e4b27eeb3194deef0b3140997e6b82f53eb7350daceb9355268009b92f70add.

UnitDeactivated

@dev Emitted when a unit deactivated. Unit is deactivated when it moved to deal @param commitmentId Commitment id @param unitId Compute unit id which deactivated Event with signature UnitDeactivated(bytes32,bytes32) and selector 0xbd9cde1bbc961036d34368ae328c38917036a98eacfb025a1ff6d2c6235d0a14.